// src/sbi/src/types/accountTypes.ts
export interface User {      
    name:   string           
    email:  string          
    phone:  string               
};

export interface Account{
    id:                 Number
    account_number:     Number
    name:               string
    email:              string
    phone:              string
    vpa:                string
    balance:            Number
    createdAt:          Date
    updatedAt:          Date
};